pyproject.toml
[project]
name = "liquidity-monitor"
version = "0.1.0"
description = "FAANG-grade global liquidity monitoring system based on Arthur Hayes' framework"
readme = "README.md"
authors = [
    { name = "gptprojectmanager", email = "gptprojectmanager@gmail.com" }
]
requires-python = ">=3.11"
license = { text = "MIT" }
keywords = ["liquidity", "central-bank", "macro", "trading", "openbb"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    # Core data platform
    "openbb>=4.0.0",
    # Time-series storage
    "questdb>=2.0.0",
    # Async HTTP client
    "httpx>=0.27.0",
    # Resilience patterns
    "tenacity>=9.0.0",
    "purgatory-circuitbreaker>=0.7.0",
    "setuptools>=70.0.0", # Required by purgatory for pkg_resources
    # Redis for pub/sub and shared state
    "redis[hiredis]>=5.0.0",
    # Metrics
    "prometheus-client>=0.21.0",
    # Configuration management
    "pydantic-settings>=2.0.0",
    "psycopg2-binary>=2.9.11",
    "pyvalet>=1.1",
    "dbnomics>=1.2.7",
    # HTML parsing for scraping fallbacks
    "beautifulsoup4>=4.12.0",
    "lxml>=5.0.0",
]

[project.optional-dependencies]
dev = [
    # Testing
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=4.1.0",

    # Linting
    "ruff>=0.4.0",

    # Type checking
    "mypy>=1.10.0",

    # Type stubs
    "types-redis>=4.6.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/liquidity"]

# Ruff configuration
[tool.ruff]
target-version = "py311"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # do not perform function calls in argument defaults
]

[tool.ruff.lint.isort]
known-first-party = ["liquidity"]

# MyPy configuration
[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_configs = true
show_error_codes = true
disable_error_code = ["import-untyped"]

[[tool.mypy.overrides]]
module = [
    "openbb",
    "openbb.*",
    "questdb",
    "questdb.*",
    "purgatory",
    "purgatory.*",
    "prometheus_client",
    "prometheus_client.*",
    "pandas",
    "pandas.*",
    "psycopg2",
    "psycopg2.*",
]
ignore_missing_imports = true

# Pytest configuration
[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
addopts = [
    "-v",
    "--tb=short",
    "--strict-markers",
]
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "e2e: End-to-end tests",
    "slow: Slow-running tests",
]

# Coverage configuration
[tool.coverage.run]
source = ["src/liquidity"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
    "if __name__ == .__main__.:",
]

[dependency-groups]
dev = [
    "pandas-stubs>=2.3.3.260113",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "types-psycopg2>=2.9.21.20251012",
]
